+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## \[Unreleased\]
 
+### Removed
+
+- Remove Visual Prompting
+  (<https://github.com/openvinotoolkit/training_extensions/pull/4291>)
+
 ## \[2.4.0\]
 
 ### New features